#3 -  added CSS to registration form, and restructured styles and pages
