+++ src/App.css
@@ -5,6 +5,14 @@
   text-align: center;
 }
 
+body {
+  background: url('assets/florian-olivo-code-unsplash.jpg') no-repeat center center fixed;
+  background-size: cover;
+  margin: 0;
+  padding: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif;
+}
+
 .logo {
   height: 6em;
   padding: 1.5em;
@@ -39,4 +47,14 @@
 
 .read-the-docs {
   color: #888;
+}
+
+.app-container {
+  background-image: url('/background.jpg');
+  background-size: cover;
+  background-position: center;
+  height: 100vh; /* Full viewport height */
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
