+++ src/styles/pages/RegistrationForm.css
@@ -1,0 +1,66 @@
+/* Ensure the body and html take full height and use Flexbox to center the form */
+
+  
+  /* Container for the registration form */
+  .registration-form-container {
+    background: white; /* White background */
+    padding: 40px; /* More padding for a spacious look */
+    border-radius: 8px; /* Slightly rounded corners */
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Subtle shadow */
+    max-width: 400px; /* Width similar to Outlook form */
+    width: 100%;
+    text-align: center; /* Center-align text inside the form */
+  }
+  
+  /* Form logo */
+  .form-logo {
+    width: 100px; /* Adjust the size as needed */
+    margin-bottom: 20px;
+  }
+  
+  /* Form title */
+  .registration-form-title {
+    font-size: 24px;
+    margin-bottom: 20px;
+    color: #333; /* Darker text color */
+  }
+  
+  /* Form fields */
+  .registration-form-field {
+    margin-bottom: 15px;
+    text-align: left; /* Left-align labels and inputs */
+  }
+  
+  /* Form labels */
+  .registration-form-field label {
+    display: block;
+    margin-bottom: 5px;
+    font-weight: bold;
+    color: #333; /* Darker text color */
+  }
+  
+  /* Form inputs */
+  .registration-form-field input {
+    width: 100%;
+    padding: 10px;
+    border: 1px solid #ccc; /* Light gray border */
+    border-radius: 4px; /* Slightly rounded corners */
+    box-sizing: border-box;
+  }
+  
+  /* Form buttons */
+  .registration-form-button {
+    width: 100%;
+    padding: 10px;
+    background-color: #0078d4; /* Outlook blue background */
+    color: white;
+    border: none;
+    border-radius: 4px; /* Slightly rounded corners */
+    cursor: pointer;
+    font-size: 16px;
+    margin-top: 20px; /* Space above the button */
+  }
+  
+  .registration-form-button:hover {
+    background-color: #005a9e; /* Darker blue on hover */
+  }
